refactor: 파일 서비스의 내부 클래스들을 추출하여 분리
