+++ src/main/java/com/nubble/backend/file/service/FileCommand.java
@@ -1,0 +1,17 @@
+package com.nubble.backend.file.service;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class FileCommand {
+
+    @Builder
+    public record FileUploadCommand(
+            String fileName,
+            byte[] fileData
+    ) {
+
+    }
+}
