+++ src/main/java/com/nubble/backend/file/service/FileInfo.java
@@ -1,0 +1,11 @@
+package com.nubble.backend.file.service;
+
+import lombok.Builder;
+
+@Builder
+public record FileInfo(
+        String fileName,
+        String baseUrl
+) {
+
+}
