+++ src/main/java/com/nubble/backend/file/service/FileUploader.java
@@ -1,0 +1,8 @@
+package com.nubble.backend.file.service;
+
+public interface FileUploader {
+
+    void upload(String fileName, byte[] file);
+
+    String getBaseUrl();
+}
